frontend: Extract NameValueTable into its own component
This patch also adds tests and ensures that the last row of the
NameValueTable doesn't have a border (which was a bug).
