+++ frontend/src/components/common/NameValueTable/index.tsx
@@ -1,0 +1,4 @@
+export * from './NameValueTable';
+export { default as NameValueTable } from './NameValueTable';
+import NameValueTable from './NameValueTable';
+export default NameValueTable;
