+++ frontend/src/components/common/index.ts
@@ -11,6 +11,8 @@ export { default as Link } from './Link';
 export * from './Loader';
 export { default as Loader } from './Loader';
 export * from './LogViewer';
+export * from './NameValueTable';
+export { default as NameValueTable } from './NameValueTable';
 export * from './ObjectEventList';
 export { default as ObjectEventList } from './ObjectEventList';
 export * from './Resource';
